requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
et-xmlfile==1.1.0
idna==3.4
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
